SERVER-49117 Remove storage validation of '$' prefixes in insert and update

+++ jstests/core/insert2.js
@@ -4,18 +4,23 @@
 //   assumes_no_implicit_collection_creation_after_drop,
 //   requires_collstats,
 //   uses_multiple_connections,
+//   requires_fcv_50,
 // ]
 
-// Create a new connection object so it won't affect the global connection when we modify
-// it's settings.
-var conn = new Mongo(db.getMongo().host);
-conn.forceWriteMode(db.getMongo().writeMode());
+var isDotsAndDollarsEnabled = db.adminCommand({getParameter: 1, featureFlagDotsAndDollars: 1})
+                                  .featureFlagDotsAndDollars.value;
+if (!isDotsAndDollarsEnabled) {
+    // Create a new connection object so it won't affect the global connection when we modify
+    // it's settings.
+    var conn = new Mongo(db.getMongo().host);
+    conn.forceWriteMode(db.getMongo().writeMode());
 
-t = conn.getDB(db.getName()).insert2;
-t.drop();
+    t = conn.getDB(db.getName()).insert2;
+    t.drop();
 
-assert.isnull(t.findOne(), "A");
-assert.writeError(t.insert({z: 1, $inc: {x: 1}}, 0, true));
-assert.isnull(t.findOne(), "B");
-// Collection should not exist
-assert.isnull(t.exists());
+    assert.isnull(t.findOne(), "A");
+    assert.writeError(t.insert({z: 1, $inc: {x: 1}}, 0, true));
+    assert.isnull(t.findOne(), "B");
+    // Collection should not exist
+    assert.isnull(t.exists());
+}
